+++ app/(authenticated)/characters/[id]/edit/types.ts
@@ -15,6 +15,7 @@ export interface Character {
   avatarUrl?: string
   defaultImageId?: string
   defaultConnectionProfileId?: string
+  npc?: boolean
   defaultImage?: {
     id: string
     filepath: string